Use trycmd for ditto-make integration tests
