+++ crates/ditto-make/tests/cmd/.ignore
@@ -1,0 +1,1 @@
+*.in
