+++ crates/ditto-make/tests/cmd/all-good/test.in/dep/ditto-src/Dep.ditto
