+++ crates/ditto-make/tests/cmd/all-good/test.in/ditto-src/A.ditto
