+++ crates/ditto-make/tests/cmd/all-good/test.in/ditto-src/B.ditto
