+++ crates/ditto-make/tests/cmd/all-good/test.stdout
@@ -54,3 +54,4 @@ build packages/dep/Util.js: js builddir/dep/Util.ast
 build packages/dep/package.json: package_json dep/ditto.toml
   description = Generating package.json for dep
 
+
