+++ crates/ditto-make/tests/cmd/duplicate-module-name/test.in/ditto-src/B.ditto
