+++ crates/ditto-make/tests/cmd/missing-module/test.in/ditto-src/A.ditto
