+++ crates/ditto-make/tests/cmd/missing-module/test.stdout
@@ -6,3 +6,4 @@ rule ast
 build builddir/A.ast builddir/A.ast-exports builddir/A.checker-warnings: ast ./ditto-src/A.ditto
   description = Checking A
 
+
