+++ crates/ditto-make/tests/cmd/module-cycle/test.in/ditto-src/A.ditto
