+++ crates/ditto-make/tests/cmd/module-cycle/test.stderr
@@ -1,0 +1,3 @@
+
+  × modules form a cycle: `A`, `B`
+
