+++ crates/ditto-make/tests/cmd/module-cycle/test.toml
@@ -1,0 +1,3 @@
+bin.name = "ditto-make-testbin"
+args = []
+fs.sandbox = true
