+++ crates/ditto-make/tests/cmd/no-codegen/test.in/build.ninja
