+++ crates/ditto-make/tests/cmd/no-codegen/test.in/dep/ditto-src/Dep.ditto
