+++ crates/ditto-make/tests/cmd/no-codegen/test.in/ditto-src/C.ditto
