+++ crates/ditto-make/tests/cmd/no-codegen/test.in/ditto-src/D.ditto
