+++ crates/ditto-make/tests/cmd/no-codegen/test.stdout
@@ -1,0 +1,21 @@
+builddir = builddir
+
+rule ast
+  command = ditto compile ast --build-dir builddir -i ${in} -o ${out}
+
+build builddir/A.ast builddir/A.ast-exports builddir/A.checker-warnings: ast ./ditto-src/A.ditto
+  description = Checking A
+
+build builddir/B.ast builddir/B.ast-exports builddir/B.checker-warnings: ast ./ditto-src/B.ditto
+  description = Checking B
+
+build builddir/C.ast builddir/C.ast-exports builddir/C.checker-warnings: ast ./ditto-src/C.ditto builddir/A.ast-exports builddir/B.ast-exports
+  description = Checking C
+
+build builddir/D.ast builddir/D.ast-exports builddir/D.checker-warnings: ast ./ditto-src/D.ditto builddir/C.ast-exports builddir/dep/Dep.ast-exports
+  description = Checking D
+
+build builddir/dep/Dep.ast builddir/dep/Dep.ast-exports: ast ./dep/ditto-src/Dep.ditto
+  description = Checking dep:Dep
+
+
