+++ crates/ditto-make/tests/cmd/self-referencing-module/test.in/ditto-src/A.ditto
