+++ crates/ditto-make/tests/cmd/self-referencing-module/test.stderr
@@ -1,0 +1,3 @@
+
+  × module `A` can't import itself!
+
