+++ crates/ditto-make/tests/cmd/target-mismatch/test.in/dep/ditto-src/Dep.ditto
