+++ crates/ditto-make/tests/cmd/target-mismatch/test.in/ditto-src/B.ditto
