+++ crates/ditto-make/tests/cmd/target-mismatch/test.stderr
@@ -1,0 +1,3 @@
+
+  × package "dep" doesn't support targets: "web"
+
