+++ crates/ditto-make/tests/cmd/unsupported-ditto-version/test.in/ditto-src/A.ditto
