+++ crates/ditto-make/tests/cmd/unsupported-ditto-version/test.stderr
@@ -1,0 +1,4 @@
+
+  × ditto version requirement not met for current_package: current version =
+  │ 0.0.0-test, wanted = ^1.0.0
+
