+++ crates/ditto-make/tests/trycmd_tests.rs
@@ -1,0 +1,4 @@
+#[test]
+fn checker_tests() {
+    trycmd::TestCases::new().case("tests/cmd/*/*.toml");
+}
